server/requirements.txt
Django==3.1.12
djongo==1.3.6
PyJWT==2.3.0
django-dotenv==1.4.2
pytz==2019.3
pymongo==3.7.2
pandas==1.0.1
statsmodels==0.11.1
scikit-learn==1.0.2
urllib3==1.26.6